move legacy C-API constants into separate files
