+++ modules/imgcodecs/include/opencv2/imgcodecs/imgcodecs_c.h
@@ -1,0 +1,1 @@
+#error "This header with legacy C API declarations has been removed from OpenCV. Legacy contants are available from legacy/constants_c.h file."
