+++ modules/photo/include/opencv2/photo/legacy/constants_c.h
@@ -1,0 +1,14 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html.
+
+#ifndef OPENCV_PHOTO_LEGACY_CONSTANTS_H
+#define OPENCV_PHOTO_LEGACY_CONSTANTS_H
+
+enum InpaintingModes
+{
+    CV_INPAINT_NS      =0,
+    CV_INPAINT_TELEA   =1
+};
+
+#endif // OPENCV_PHOTO_LEGACY_CONSTANTS_H
