+++ modules/video/include/opencv2/video/tracking.hpp
@@ -47,14 +47,6 @@
 #include "opencv2/core.hpp"
 #include "opencv2/imgproc.hpp"
 
-enum
-{
-    CV_LKFLOW_PYR_A_READY = 1,
-    CV_LKFLOW_PYR_B_READY = 2,
-    CV_LKFLOW_INITIAL_GUESSES = 4,
-    CV_LKFLOW_GET_MIN_EIGENVALS = 8
-};
-
 namespace cv
 {
