+++ modules/videoio/src/cap_gphoto2.cpp
@@ -1005,7 +1005,7 @@ void DigitalCameraCapture::readFrameFromFile(CameraFile * file, OutputArray outp
         Mat buf = Mat(1, size, CV_8UC1, (void *) data);
         if(!buf.empty())
         {
-            frame = imdecode(buf, CV_LOAD_IMAGE_UNCHANGED);
+            frame = imdecode(buf, IMREAD_UNCHANGED);
         }
         frame.copyTo(outputFrame);
     }
